requirements.txt
psycopg2==2.8.3
python-telegram-bot==11.1.0
requests==2.22.0
